# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_AVGsync_keyfreestay_backend_internal_model_request.CreatePaymentMethodRequest:
    properties:
      card_brand:
        description: Платёжная система карты.
        enum:
        - visa
        - mastercard
        - mir
        - unionpay
        - other
        example: visa
        type: string
      card_last4:
        description: Последние 4 цифры карты. Полный номер карты API не принимает.
        example: "4242"
        type: string
      cardholder_name:
        description: Имя держателя карты.
        example: IVAN PETROV
        maxLength: 100
        minLength: 2
        type: string
      expiry_month:
        description: Месяц окончания срока действия карты.
        example: 12
        maximum: 12
        minimum: 1
        type: integer
      expiry_year:
        description: Год окончания срока действия карты.
        example: 2028
        maximum: 2099
        minimum: 2024
        type: integer
      is_default:
        description: Сделать метод основным для пользователя.
        example: true
        type: boolean
    required:
    - card_brand
    - card_last4
    - cardholder_name
    - expiry_month
    - expiry_year
    type: object
  github_com_AVGsync_keyfreestay_backend_internal_model_request.LoginRequest:
    properties:
      email:
        description: Email зарегистрированного пользователя.
        example: ivan.petrov@example.com
        type: string
      password:
        description: Пароль пользователя.
        example: Str0ngPass!2026
        type: string
    required:
    - email
    - password
    type: object
  github_com_AVGsync_keyfreestay_backend_internal_model_request.NewUserRequest:
    properties:
      email:
        description: Уникальный email, используется для входа.
        example: ivan.petrov@example.com
        type: string
      full_name:
        description: Полное имя пользователя, отображается в профиле.
        example: Иван Петров
        maxLength: 64
        minLength: 3
        type: string
      password:
        description: Пароль, минимум 8 символов.
        example: Str0ngPass!2026
        minLength: 8
        type: string
      phone:
        description: Номер телефона в международном формате E.164.
        example: "+1234567890"
        type: string
    required:
    - email
    - full_name
    - password
    type: object
...
  github_com_AVGsync_keyfreestay_backend_internal_model_request.UserUpdateRequest:
    properties:
      email:
        description: Новый email. Если поле не передано, текущее значение не меняется.
        example: ivan.new@example.com
        type: string
      full_name:
        description: Новое полное имя. Если поле не передано, текущее значение не
          меняется.
        example: Иван Сергеевич Петров
        maxLength: 64
        minLength: 3
        type: string
      phone:
        description: Новый номер телефона. Если поле не передано, текущее значение
          не меняется.
        example: "+1234567890"
        type: string
    type: object
...
  github_com_AVGsync_keyfreestay_backend_internal_model_response.PaymentMethodListResponse:
    properties:
      items:
        description: Платёжные методы пользователя.
        items:
          $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_response.PaymentMethodResponse'
        type: array
      total:
        description: Общее количество методов в ответе.
        example: 2
        type: integer
    type: object
  github_com_AVGsync_keyfreestay_backend_internal_model_response.PaymentMethodResponse:
    properties:
      card_brand:
        description: Платёжная система карты.
        enum:
        - visa
        - mastercard
        - mir
        - unionpay
        - other
        example: visa
        type: string
      card_last4:
        description: Последние 4 цифры карты.
        example: "4242"
        type: string
      cardholder_name:
        description: Имя держателя карты.
        example: IVAN PETROV
        type: string
      created_at:
        description: Дата добавления платёжного метода.
        example: "2026-05-15T12:00:00Z"
        type: string
      expiry_month:
        description: Месяц окончания срока действия карты.
        example: 12
        maximum: 12
        minimum: 1
        type: integer
      expiry_year:
        description: Год окончания срока действия карты.
        example: 2028
        maximum: 2099
        minimum: 2024
        type: integer
      id:
        description: UUID платёжного метода.
        example: 660e8400-e29b-41d4-a716-446655440000
        type: string
      is_default:
        description: Является ли метод основным.
        example: true
        type: boolean
    type: object
  github_com_AVGsync_keyfreestay_backend_internal_model_response.UserResponse:
    properties:
      email:
        description: Email пользователя.
        example: ivan.petrov@example.com
        type: string
      full_name:
        description: Полное имя пользователя.
        example: Иван Петров
        type: string
      id:
        description: UUID пользователя.
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      phone:
        description: Телефон пользователя, если указан.
        example: "+1234567890"
        type: string
      role:
        description: Роль пользователя.
        enum:
        - user
        - admin
        example: user
        type: string
      subscription_expires:
        description: Дата окончания подписки. Для бесплатного тарифа может быть null.
        example: "2026-12-31T23:59:59Z"
        type: string
      subscription_plan:
        description: Тарифный план пользователя.
        enum:
        - free
        - pro
        - enterprise
        example: free
        type: string
    type: object
info:
  contact:
    email: support@keyfreestay.local
    name: KeyFreeStay API Support
  description: |-
    REST API сервиса KeyFreeStay для краткосрочной аренды жилья без ключей.

    API покрывает регистрацию и вход, профиль пользователя, объекты жилья, изображения жилья, бронирования, платёжные методы и публичную контактную заявку.
    Схема авторизации: вызовите `POST /auth/login`; API установит JWT в HttpOnly cookie `token`. Защищённые маршруты читают JWT из cookie.
    Swagger UI доступен только при `DEBUG=true`: `/swagger`, JSON спецификация — `/swagger/doc.json`.
  title: KeyFreeStay API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: |-
        Проверяет email и пароль пользователя.
        При успехе устанавливает JWT в HttpOnly cookie `token` на 24 часа.
        Защищённые маршруты ожидают эту cookie в заголовке `Cookie`.
      parameters:
      - description: Email и пароль зарегистрированного пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_request.LoginRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: Пустой ответ со статусом 200; JWT передан в Set-Cookie
          headers:
            Set-Cookie:
              description: token=<jwt>; HttpOnly; Secure; SameSite=Strict; Path=/;
                Max-Age=86400
              type: string
          schema:
            type: string
        "400":
          description: invalid request body — тело запроса не JSON или структура не
            совпадает со схемой
          schema:
            type: string
        "401":
          description: invalid email or password — неверный email или пароль
          schema:
            type: string
      summary: Войти в аккаунт
      tags:
      - Авторизация
  /auth/register:
    post:
      consumes:
      - application/json
      description: |-
        Создаёт аккаунт гостя по полному имени, email, телефону и паролю.
        Email должен быть уникальным без учёта регистра. Пароль хешируется и никогда не возвращается в ответе.
        Новый пользователь получает роль `user` и тариф `free`. После регистрации нужно вызвать /auth/login, чтобы получить cookie `token`.
      parameters:
      - description: Данные регистрации пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_request.NewUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Профиль созданного пользователя
          schema:
            $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_response.UserResponse'
        "400":
          description: invalid request body — тело запроса не JSON или структура не
            совпадает со схемой
          schema:
            type: string
        "500":
          description: failed to register new user — пользователь не создан
          schema:
            type: string
      summary: Зарегистрировать нового пользователя
      tags:
      - Авторизация
...
  /me:
    get:
      description: |-
        Возвращает профиль пользователя из JWT claims в cookie `token`.
        Используйте этот метод для проверки текущей сессии и отображения данных профиля.
      produces:
      - application/json
      responses:
        "200":
          description: Профиль текущего пользователя
          schema:
            $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_response.UserResponse'
        "401":
          description: unauthorized — cookie `token` отсутствует, просрочена или недействительна
          schema:
            type: string
        "500":
          description: claims not found in context / failed to get user — внутренняя
            ошибка получения профиля
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Получить профиль текущего пользователя
      tags:
      - Пользователи
    patch:
      consumes:
      - application/json
      description: |-
        Частично обновляет email, полное имя и/или телефон текущего пользователя.
        Все поля необязательные. Не переданные поля сохраняют прежнее значение.
        Если email меняется, он должен оставаться уникальным.
      parameters:
      - description: Поля профиля, которые нужно изменить
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_request.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённый профиль пользователя
          schema:
            $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_response.UserResponse'
        "400":
          description: invalid request body — тело запроса не JSON или структура не
            совпадает со схемой
          schema:
            type: string
        "401":
          description: unauthorized — cookie `token` отсутствует, просрочена или недействительна
          schema:
            type: string
        "500":
          description: claims not found in context / failed to update user — внутренняя
            ошибка обновления профиля
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Обновить профиль текущего пользователя
      tags:
      - Пользователи
  /payments:
    get:
      description: |-
        Возвращает платёжные методы текущего пользователя.
        Список отсортирован так, что основной метод (`is_default=true`) идёт первым, затем более новые методы.
      produces:
      - application/json
      responses:
        "200":
          description: Список платёжных методов пользователя
          schema:
            $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_response.PaymentMethodListResponse'
        "401":
          description: unauthorized — cookie `token` отсутствует, просрочена или недействительна
          schema:
            type: string
        "500":
          description: failed to get payment method list — список не получен
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Получить список платёжных методов
      tags:
      - Платёжные методы
  /payments/{id}:
    delete:
      description: |-
        Удаляет платёжный метод текущего пользователя по ID.
        Пользователь может удалить только свой платёжный метод.
      parameters:
      - description: UUID платёжного метода
        example: 550e8400-e29b-41d4-a716-446655440000
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Пустой ответ со статусом 200
          schema:
            type: string
        "400":
          description: missing payment method ID — id не передан в path
          schema:
            type: string
        "401":
          description: unauthorized — cookie `token` отсутствует, просрочена или недействительна
          schema:
            type: string
        "500":
          description: failed to delete payment method — платёжный метод не найден,
            не принадлежит пользователю или не удалён
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Удалить платёжный метод
      tags:
      - Платёжные методы
  /payments/create:
    put:
      consumes:
      - application/json
      description: |-
        Создаёт платёжный метод текущего пользователя.
        API хранит только безопасные карточные данные: последние 4 цифры, бренд, срок действия, имя держателя и флаг основного метода.
        Если `is_default=true`, текущий основной метод пользователя сбрасывается.
      parameters:
      - description: Данные платёжного метода
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_request.CreatePaymentMethodRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Созданный платёжный метод
          schema:
            $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_response.PaymentMethodResponse'
        "400":
          description: invalid request body — тело запроса не JSON или структура не
            совпадает со схемой
          schema:
            type: string
        "401":
          description: unauthorized — cookie `token` отсутствует, просрочена или недействительна
          schema:
            type: string
        "500":
          description: failed to create payment method — платёжный метод не создан
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Добавить платёжный метод
      tags:
      - Платёжные методы
...
schemes:
- http
securityDefinitions:
  CookieAuth:
    description: 'Cookie авторизации. Формат заголовка: Cookie: token={jwt}'
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"