# backend/internal/transport/http/handler/user.go
package handler

import (
	"context"
	"encoding/json"
	"net/http"

	"github.com/AVGsync/keyfreestay/backend/internal/model"
	"github.com/AVGsync/keyfreestay/backend/internal/model/request"
	"github.com/AVGsync/keyfreestay/backend/internal/model/response"
)

type UserUseCase interface {
	RegisterNewUser(user *request.NewUserRequest, ctx context.Context) (response.UserResponse, error)
	AuthenticateUser(req request.LoginRequest, ctx context.Context) (string, error)
	GetUserByID(id string, ctx context.Context) (response.UserResponse, error)
	UpdateUser(user *request.UserUpdateRequest, ctx context.Context) (response.UserResponse, error)
}

type UserHandler struct {
	useCase UserUseCase
}

func NewUserHandler(useCase UserUseCase) *UserHandler {
	return &UserHandler{
		useCase: useCase,
	}
}

// RegisterNewUser godoc
//
// @Summary Зарегистрировать нового пользователя
// @Description Создаёт аккаунт гостя по полному имени, email, телефону и паролю.
// @Description Email должен быть уникальным без учёта регистра. Пароль хешируется и никогда не возвращается в ответе.
// @Description Новый пользователь получает роль `user` и тариф `free`. После регистрации нужно вызвать /auth/login, чтобы получить cookie `token`.
// @Tags Авторизация
// @Accept json
// @Produce json
// @Param request body request.NewUserRequest true "Данные регистрации пользователя"
// @Success 200 {object} response.UserResponse "Профиль созданного пользователя"
// @Failure 400 {string} string "invalid request body — тело запроса не JSON или структура не совпадает со схемой"
// @Failure 500 {string} string "failed to register new user — пользователь не создан"
// @Router /auth/register [post]
func (h *UserHandler) RegisterNewUser() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		u := &request.NewUserRequest{}
		if err := json.NewDecoder(r.Body).Decode(u); err != nil {
			http.Error(w, "invalid request body", http.StatusBadRequest)
			return
		}

		user, err := h.useCase.RegisterNewUser(u, r.Context())
		if err != nil {
			http.Error(w, "failed to register new user", http.StatusInternalServerError)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		json.NewEncoder(w).Encode(user)
	}
}

// LoginUser godoc
//
// @Summary Войти в аккаунт
// @Description Проверяет email и пароль пользователя.
// @Description При успехе устанавливает JWT в HttpOnly cookie `token` на 24 часа.
// @Description Защищённые маршруты ожидают эту cookie в заголовке `Cookie`.
// @Tags Авторизация
// @Accept json
// @Produce plain
// @Param request body request.LoginRequest true "Email и пароль зарегистрированного пользователя"
// @Success 200 {string} string "Пустой ответ со статусом 200; JWT передан в Set-Cookie"
// @Header 200 {string} Set-Cookie "token=<jwt>; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=86400"
// @Failure 400 {string} string "invalid request body — тело запроса не JSON или структура не совпадает со схемой"
// @Failure 401 {string} string "invalid email or password — неверный email или пароль"
// @Router /auth/login [post]
func (h *UserHandler) LoginUser() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		u := &request.LoginRequest{}
		if err := json.NewDecoder(r.Body).Decode(u); err != nil {
			http.Error(w, "invalid request body", http.StatusBadRequest)
			return
		}

		token, err := h.useCase.AuthenticateUser(*u, r.Context())
		if err != nil {
			http.Error(w, "invalid email or password", http.StatusUnauthorized)
			return
		}

		http.SetCookie(w, &http.Cookie{
			Name:     "token",
			Value:    token,
			HttpOnly: true,
			Secure:   true,
			SameSite: http.SameSiteStrictMode,
			Path:     "/",
			MaxAge:   60 * 60 * 24,
		})

		w.WriteHeader(http.StatusOK)
	}
}

// GetMe godoc
//
// @Summary Получить профиль текущего пользователя
// @Description Возвращает профиль пользователя из JWT claims в cookie `token`.
// @Description Используйте этот метод для проверки текущей сессии и отображения данных профиля.
// @Tags Пользователи
// @Produce json
// @Security CookieAuth
// @Success 200 {object} response.UserResponse "Профиль текущего пользователя"
// @Failure 401 {string} string "unauthorized — cookie `token` отсутствует, просрочена или недействительна"
// @Failure 500 {string} string "claims not found in context / failed to get user — внутренняя ошибка получения профиля"
// @Router /me [get]
func (h *UserHandler) GetMe() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		claims, ok := r.Context().Value("claims").(*model.Claims)
		if !ok {
			http.Error(w, "claims not found in context", http.StatusInternalServerError)
			return
		}

		user, err := h.useCase.GetUserByID(claims.UserID, r.Context())
		if err != nil {
			http.Error(w, "failed to get user", http.StatusInternalServerError)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		json.NewEncoder(w).Encode(user)
	}
}

// UpdateUser godoc
//
// @Summary Обновить профиль текущего пользователя
// @Description Частично обновляет email, полное имя и/или телефон текущего пользователя.
// @Description Все поля необязательные. Не переданные поля сохраняют прежнее значение.
// @Description Если email меняется, он должен оставаться уникальным.
// @Tags Пользователи
// @Accept json
// @Produce json
// @Security CookieAuth
// @Param request body request.UserUpdateRequest true "Поля профиля, которые нужно изменить"
// @Success 200 {object} response.UserResponse "Обновлённый профиль пользователя"
// @Failure 400 {string} string "invalid request body — тело запроса не JSON или структура не совпадает со схемой"
// @Failure 401 {string} string "unauthorized — cookie `token` отсутствует, просрочена или недействительна"
// @Failure 500 {string} string "claims not found in context / failed to update user — внутренняя ошибка обновления профиля"
// @Router /me [patch]
func (h *UserHandler) UpdateUser() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		_, ok := r.Context().Value("claims").(*model.Claims)
		if !ok {
			http.Error(w, "claims not found in context", http.StatusInternalServerError)
			return
		}

		u := &request.UserUpdateRequest{}
		if err := json.NewDecoder(r.Body).Decode(u); err != nil {
			http.Error(w, "invalid request body", http.StatusBadRequest)
			return
		}

		user, err := h.useCase.UpdateUser(u, r.Context())
		if err != nil {
			http.Error(w, "failed to update user", http.StatusInternalServerError)
			return
		}

		w.Header().Set("Content-Type", "application/json")
		json.NewEncoder(w).Encode(user)
	}
}